Move the in-flight invalidation tests into CacheTests
The new tests are ordinary Cache tests and belong next to the existing
invalidation and concurrency cases rather than in a file of their own.
